split tests, init rust tests, init rust badge
