+++ src/keep_cool.rs
@@ -1,13 +1,14 @@
 use pyo3::prelude::*;
 use pyo3::types::{PyList, PyModule};
 use std::fs::File;
-use std::io::{BufRead, BufReader, Write};
-use flate2::read::GzDecoder;
+use std::io::{Write};
 use rayon::prelude::*;
 use rayon::ThreadPoolBuilder;
 use std::cmp::Ordering;
 use sprs::{CsMat, TriMat};
 use sprs::io::write_matrix_market;
+use crate::types::{Region, CoolRegion};
+use crate::reader::{read_meth, parse_chromsizes, parse_region};
 
 #[allow(clippy::too_many_arguments)]
 #[pyfunction]
@@ -92,7 +93,7 @@ pub fn parse_cools(
         coolfiles
             .par_iter()
             .map(|coolfile| {
-                let coolregions = parse_cool(coolfile);
+                let coolregions = read_meth(coolfile);
                 parsed_regions
                     .par_iter()
                     .map(|region| {
@@ -222,159 +223,4 @@ fn tupvec_to_sparse(dense: Vec<Vec<(f32, f32, f32)>>) -> (CsMat<f32>, CsMat<f32>
         }
     }
     (mat1.to_csr(), mat2.to_csr(), mat3.to_csr())
-}
-
-fn parse_cool(_f: &str) -> Vec<CoolRegion> {
-    let mut coolregions: Vec<CoolRegion> = Vec::new();
-    let reader = BufReader::new(GzDecoder::new(File::open(_f).unwrap()));
-    for line in reader.lines() {
-        match line {
-            Ok(line) => {
-                let fields: Vec<&str> = line.split('\t').collect();
-                let chrom = fields[0].to_string();
-                let pos = fields[1].parse::<u32>().unwrap();
-                let meth = fields[4].parse::<u32>().unwrap();
-                let total = fields[5].parse::<u32>().unwrap();
-                coolregions.push(
-                    CoolRegion{
-                        chrom,
-                        pos,
-                        meth,
-                        total,
-                    }
-                );
-            }
-            Err(_e) => {
-                panic!("Error reading file {}", _f);
-            }
-        }
-    }
-    coolregions
-}
-
-fn parse_region(reg: String, class: String) -> Vec<Region> {
-    let mut regions = Vec::new();
-    let sample = reg.clone();
-
-    // Get suffix from reg
-    let suffix = reg.split('.').next_back().unwrap();
-    // Two options: gz (bed.gz), bed(bed)
-    let reader: Box<dyn BufRead> = match suffix {
-        "gz" => {
-            println!("Region parse match gz");
-            Box::new(BufReader::new(GzDecoder::new(File::open(reg).unwrap())))
-        },
-        "bed" => {
-            Box::new(BufReader::new(File::open(reg).unwrap()))
-        },
-        _ => panic!("File format not supported"),
-    };
-
-    let lines = reader.lines();
-    for line in lines {
-        match line {
-            Ok(line) => {
-                let fields: Vec<&str> = line.split('\t').collect();
-                let chrom = fields[0].to_string();
-                let start = fields[1].to_string();
-                let end = fields[2].to_string();
-                let name: String = if fields.len() > 3 {
-                    fields[3].to_string()
-                } else {
-                    format!("{}:{}-{}", chrom, start, end)
-                };
-                // check if start, end have commas
-                if start.contains(",") {
-                    let start: Vec<u32> = start.split(',').map(|x| x.parse::<u32>().unwrap()).collect();
-                    let end: Vec<u32> = end.split(',').map(|x| x.parse::<u32>().unwrap()).collect();
-                    regions.push(
-                        Region{
-                            chrom,
-                            start,
-                            end,
-                            name,
-                            class: class.to_string()
-                        }
-                    );
-                } else {
-                    let start = start.parse::<u32>().unwrap();
-                    let end = end.parse::<u32>().unwrap();
-                    regions.push(
-                        Region{
-                            chrom,
-                            start: vec![start],
-                            end: vec![end],
-                            name,
-                            class: class.to_string()
-                        }
-                    );
-                }
-
-            }
-            Err(_e) => {
-                panic!("Error reading file {}", sample);
-            }
-        }
-    }
-    regions
-}
-
-fn parse_chromsizes(file: &str, binsize: u32) -> Vec<Region> {
-    let mut regions: Vec<Region> = Vec::new();
-    let reader = BufReader::new(File::open(file).unwrap());
-    for line in reader.lines() {
-        match line {
-            Ok(line) => {
-                let fields: Vec<&str> = line.split('\t').collect();
-                let chrom = fields[0].to_string();
-                let chromsize = fields[1].parse::<u32>().unwrap();
-                let mut start = 0;
-                let mut end = start + binsize;
-                while end < chromsize {
-                    regions.push(
-                        Region{
-                            chrom: chrom.clone(),
-                            start: vec![start],
-                            end: vec![end],
-                            name: format!("{}:{}-{}", chrom, start, end),
-                            class: "bin".to_string(),
-                        }
-                    );
-                    start = end;
-                    end += binsize;
-                    // Capture chromosome end
-                    if end >= chromsize {
-                        regions.push(
-                            Region{
-                                chrom: chrom.clone(),
-                                start: vec![start],
-                                end: vec![chromsize],
-                                name: format!("{}:{}-{}", chrom, start, end),
-                                class: "bin".to_string(),
-                            }
-                        );
-                    }
-                }
-            },
-            Err(_e) => {
-                panic!("Error reading file {}", file);
-            }
-        }
-    }
-    regions
-}
-
-struct Region {
-    chrom: String,
-    start: Vec<u32>,
-    end: Vec<u32>,
-    name: String,
-    class: String,
-}
-
-struct CoolRegion {
-    chrom: String,
-    pos: u32,
-    meth: u32,
-    total: u32,
 }
