+++ src/lib.rs
@@ -1,5 +1,9 @@
 use pyo3::prelude::*;
 mod keep_cool;
+mod reader;
+mod types;
+#[cfg(test)]
+mod tests;
 
 #[pymodule]
 fn linkapy(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
