+++ src/tests/data/region.bed
@@ -1,0 +1,1 @@
+chr1	100	200
